kratos-dashboard/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyodbc==5.0.1
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic[dotenv]==2.5.0
python-jose[cryptography]==3.3.0
